mine_labeling_project/requirements.txt
# Mine Labeling Project - Python Dependencies

# Core Scientific Computing
numpy>=1.21.0
scipy>=1.7.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.5.0
scikit-image>=0.19.0

# XTF Processing
pyxtf>=1.0.0

# Data Processing
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration
PyYAML>=6.0
python-dotenv>=0.19.0

# Data Augmentation
albumentations>=1.1.0
imgaug>=0.4.0

# CLI and Progress
click>=8.0.0
tqdm>=4.62.0

# Utilities
lxml>=4.6.0  # For XML parsing

# Development (optional)
pytest>=6.2.0
black>=21.0
flake8>=3.9.0
